// application/typescript/packages/@aws-prototype/website/src/api/types.ts
/**********************************************************************************************************************
Licensed to the Apache Software Foundation (ASF) under one or more contributor license agreements.  See the NOTICE file
distributed with this work for additional information regarding copyright ownership.  The ASF licenses this file
to you under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance
with the License.  You may obtain a copy of the License at

  http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an
"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the License for the
specific language governing permissions and limitations under the License.   
 **********************************************************************************************************************/
import { QueryKey } from 'react-query';

export interface APIRequest {
  apiName: string;
  queryKey: QueryKey;
  reactQueryOptions?: any;
  apiOptions?: any;
  path?: string;
  getPath?: (payload: any) => string;
}

export interface Subscription<TPayload extends unknown, TData extends unknown> {
  type: string;
  payload: TPayload;
  queryKeys?: QueryKey[];
  onDataReceived?: (data: TData) => void;
}

export interface WebSocketRequest {
  name: string;
  toSendHeartbeat?: boolean;
}